+++ klause/src/commonMain/kotlin/com/eignex/klause/factor/global/ValuePrecedeInvariant.kt
@@ -6,11 +6,11 @@ import com.eignex.klause.localsearch.LocalSearchState
 import com.eignex.klause.localsearch.MoveSink
 
 /** LS invariant logic for `value_precede`. */
-internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, private val xs: IntArray) : Invariant {
+internal class ValuePrecedeInvariant(private val s: Long, private val t: Long, private val xs: IntArray) : Invariant {
 
     override fun isViolated(state: LocalSearchState, factorId: Int): Boolean {
         val i = firstStOccurrence(state)
-        return i >= 0 && state.assignment.intValue(xs[i]) == t.toLong()
+        return i >= 0 && state.assignment.intValue(xs[i]) == t
     }
 
     override fun violationDegree(state: LocalSearchState, factorId: Int): Int =
@@ -32,25 +32,25 @@ internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, pri
         var firstBadT = -1
         for (i in xs.indices) {
             val v = state.assignment.intValue(xs[i])
-            if (v == s.toLong()) {
+            if (v == s) {
                 firstS = i
                 break
             }
-            if (v == t.toLong() && firstBadT < 0) firstBadT = i
+            if (v == t && firstBadT < 0) firstBadT = i
         }
         if (firstBadT < 0 || firstBadT > firstS) return
         val badVar = xs[firstBadT]
         val badDom = state.problem.intDomains[badVar]
         for (cand in longArrayOf(badDom.min, badDom.max)) {
-            if (cand != t.toLong() && cand in badDom) {
+            if (cand != t && cand in badDom) {
                 sink.addChannelingIntSet(state, badVar, cand)
                 break
             }
         }
         for (i in 0..firstBadT) {
             val v = xs[i]
-            if (s.toLong() in state.problem.intDomains[v] && state.assignment.intValue(v) != s.toLong()) {
-                sink.addChannelingIntSet(state, v, s.toLong())
+            if (s in state.problem.intDomains[v] && state.assignment.intValue(v) != s) {
+                sink.addChannelingIntSet(state, v, s)
                 break
             }
         }
@@ -68,7 +68,7 @@ internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, pri
             var pick = -1L
             var seen = 0
             d.forEach { w ->
-                if (w != v && (w == s.toLong() || (w != t.toLong() && v != s.toLong()))) {
+                if (w != v && (w == s || (w != t && v != s))) {
                     seen++
                     if (state.rng.nextInt(seen) == 0) pick = w
                 }
@@ -83,12 +83,12 @@ internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, pri
         for (i in xs.indices) {
             val v = xs[i]
             if (state.assumptions.isFrozenInt(v)) {
-                if (state.assignment.intValue(v) == t.toLong()) return false
+                if (state.assignment.intValue(v) == t) return false
                 continue
             }
             val d = state.problem.intDomains[v]
             var pick = -1L
-            d.forEach { if (pick < 0 && it != t.toLong()) pick = it }
+            d.forEach { if (pick < 0 && it != t) pick = it }
             if (pick < 0) return false
             state.assignment.setInt(v, pick)
         }
@@ -98,7 +98,7 @@ internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, pri
     private fun firstStOccurrence(state: LocalSearchState): Int {
         for (i in xs.indices) {
             val v = state.assignment.intValue(xs[i])
-            if (v == s.toLong() || v == t.toLong()) return i
+            if (v == s || v == t) return i
         }
         return -1
     }
@@ -107,8 +107,8 @@ internal class ValuePrecedeInvariant(private val s: Int, private val t: Int, pri
         var count = 0
         for (i in xs.indices) {
             val v = if (xs[i] == intVar) newValue else state.assignment.intValue(xs[i])
-            if (v == s.toLong()) return count
-            if (v == t.toLong()) count++
+            if (v == s) return count
+            if (v == t) count++
         }
         return count
     }
